[oiSH/ogc] Changed to be more future proof; including new shader types; compute, tesselation, raytracing and mesh/task shader types [app] Added simple compute shader example

+++ ogc/src/graphics/gl/vulkan/vktexture.cpp
@@ -242,7 +242,7 @@ bool Texture::getPixelsGpu(Vec2u start, Vec2u length, CopyBuffer &output) {
 	//Copy buffer to copy buffer
 	void *memory;
 	vkCheck<0x18, VkTexture>(vkMapMemory(graphics.device, dstMemory, 0, VK_WHOLE_SIZE, 0, &memory), "Couldn't map intermediate memory");
-	output = CopyBuffer((u8*)memory, requirements.size);
+	output = CopyBuffer((u8*)memory, (u32) requirements.size);
 	vkUnmapMemory(graphics.device, dstMemory);
 
 	//Destroy buffer
